pyproject.toml
[tool.poetry]
name = "rcos-io"
version = "0.1.0"
description = ""
authors = ["Frank Matranga <thefrankmatranga@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "rcos_io"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
gunicorn = "^20.1.0"
celery = "^5.2.7"
python-dotenv = "^1.0.0"
sentry-sdk = "^1.24.0"
whitenoise = "^6.4.0"
markdownify = "^0.11.6"
django-markdownify = "^0.9.3"
django-magiclink = "^1.2.0"
django-anymail = "^10.0"
django-crispy-forms = "^2.0"
crispy-bulma = "^0.9.0"
django-debug-toolbar = "^4.1.0"
gql = "^3.4.1"
requests-toolbelt = "^1.0.0"
psycopg2-binary = "^2.9.6"
redis = "^4.5.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.272"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "W", "F", "I", "UP", "RUF"]